updates viewer dependency to network-viewer v1.0.0
