+++ src/pypowsybl_jupyter/util.py
@@ -1,0 +1,22 @@
+# Copyright (c) 2024, RTE (http://www.rte-france.com)
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this
+# file, You can obtain one at http://mozilla.org/MPL/2.0/.
+# SPDX-License-Identifier: MPL-2.0
+#
+
+def _get_svg_string(svg) -> str:
+    if isinstance(svg, str):
+        return svg
+    elif hasattr(svg, '_repr_svg_'):
+        return svg._repr_svg_()
+    else:
+        raise ValueError('svg argument should be a string or provide a _repr_svg_ method.')
+
+def _get_svg_metadata(svg) -> str:
+    if isinstance(svg, str):
+        return None
+    elif hasattr(svg, '_metadata'):
+        return svg._metadata
+    else:
+        raise ValueError('svg argument provide a _metadata method.')
